requirements.txt
# ===== Core =====
numpy>=1.26.0
pandas>=2.1.0
pyarrow>=14.0.0
scikit-learn>=1.3.0
Pillow>=10.0.0
opencv-python>=4.8.0
tqdm>=4.66.0

# ===== FAISS (indexing engine) =====
faiss-cpu>=1.8.0
# If the user has GPU, they can manually install: faiss-gpu>=1.8.0

# ===== Deep Learning =====
torch>=2.2.0
torchvision>=0.17.0
torchaudio>=2.2.0
transformers>=4.40.2
timm>=1.0.7

# ===== Models =====
ultralytics>=8.1.0   # YOLOv8 for object detection
# BLIP is included via HuggingFace Transformers

# ===== Web App ====
streamlit>=1.33.0

# ===== Evaluation & Visualization ====
matplotlib>=3.7.0
seaborn>=0.12.0
tabulate>=0.9.0
